DisciplinaService e UsuarioDisciplinaService

+++ src/main/java/br/com/acerta/service/DisciplinaService.java
@@ -3,19 +3,22 @@ package br.com.acerta.service;
 import br.com.acerta.dto.DisciplinaDto;
 import br.com.acerta.model.Disciplina;
 import br.com.acerta.repository.DisciplinaRepository;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class DisciplinaService {
+
     @Autowired
-    private DisciplinaRepository disciplinaRepository;
+    DisciplinaRepository disciplinaRepository;
 
     public Disciplina criarDisciplina(DisciplinaDto disciplinaDto) {
-        Disciplina disciplina = new Disciplina();
-        disciplina.setNome(disciplinaDto.getNome());
+        var disciplina = new Disciplina();
+        BeanUtils.copyProperties(disciplinaDto, disciplina);
         return disciplinaRepository.save(disciplina);
     }
 
@@ -24,18 +27,30 @@ public class DisciplinaService {
     }
 
     public Disciplina editarDisciplina(Long id, DisciplinaDto disciplinaDto) {
-        Disciplina disciplina = disciplinaRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Disciplina não encontrada"));
-        disciplina.setNome(disciplinaDto.getNome());
-        return disciplinaRepository.save(disciplina);
+        Optional<Disciplina> disciplina = disciplinaRepository.findById(id);
+
+        if (disciplina.isEmpty()) {
+            return null;
+        }
+
+        BeanUtils.copyProperties(disciplinaDto, disciplina.get());
+        return disciplinaRepository.save(disciplina.get());
     }
 
     public void excluirDisciplina(Long id) {
+        Optional<Disciplina> disciplina = disciplinaRepository.findById(id);
+        if (disciplina.isEmpty()) {
+            return;
+        }
+
         disciplinaRepository.deleteById(id);
     }
 
     public Disciplina buscarDisciplinaPorId(Long id) {
-        return disciplinaRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Disciplina não encontrada"));
+        Optional<Disciplina> disciplina = disciplinaRepository.findById(id);
+        if (disciplina.isEmpty()) {
+            return null;
+        }
+        return disciplina.get();
     }
 }
